Rename files that aren't base classes any more
Now that we can easily see which classes are and aren't supposed to be
Base classes, we can change the names of those that are used directly instead
of creating a derived class.
