+++ src/wxui/menucombobox.cpp
@@ -14,26 +14,26 @@
 
 using namespace GenEnum;
 
-#include "menuspin_base.h"
+#include "menucombobox.h"
 
 namespace wxue_img
 {
-    extern const unsigned char spin_ctrl_double_png[219];
-    extern const unsigned char spin_ctrl_png[300];
-    extern const unsigned char spinbtn_png[192];
+    extern const unsigned char bmp_combo_box_png[492];
+    extern const unsigned char wxChoice_png[330];
+    extern const unsigned char wxComboBox_png[233];
 }
 
-MenuSpin::MenuSpin() : wxMenu()
+MenuCombobox::MenuCombobox() : wxMenu()
 {
     if (!wxImage::FindHandler(wxBITMAP_TYPE_PNG))
         wxImage::AddHandler(new wxPNGHandler);
 
-    auto* menu_item = Append(gen_wxSpinCtrl, "Insert wxSpinCtrl");
-    menu_item->SetBitmap(wxue_img::bundle_spin_ctrl_png());
-    auto* menu_item_2 = Append(gen_wxSpinCtrlDouble, "Insert wxSpinCtrlDouble");
-    menu_item_2->SetBitmap(wxue_img::bundle_spin_ctrl_double_png());
-    auto* menu_item_3 = Append(gen_wxSpinButton, "Insert wxSpinButton");
-    menu_item_3->SetBitmap(wxue_img::bundle_spinbtn_png());
+    auto* menu_item = Append(gen_wxComboBox, "Insert wxComboBox");
+    menu_item->SetBitmap(wxue_img::bundle_wxComboBox_png());
+    auto* menu_item_2 = Append(gen_wxChoice, "Insert wxChoice");
+    menu_item_2->SetBitmap(wxue_img::bundle_wxChoice_png());
+    auto* menu_item_3 = Append(gen_wxBitmapComboBox, "Insert wxBitmapComboBox");
+    menu_item_3->SetBitmap(wxue_img::bundle_bmp_combo_box_png());
 }
 
 // ************* End of generated code ***********
