+++ src/wxui/wxui_code.cmake
@@ -4,24 +4,49 @@
 
 set (wxue_generated_code
 
+    # Non-base classes
     ${CMAKE_CURRENT_LIST_DIR}/../customprops/include_files_dlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/code_compare.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/../internal/convert_img_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/debugsettings.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/../internal/msgframe_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/node_info.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/node_search_dlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/undo_info.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/unused_gen_dlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../internal/xrcpreview.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../tools/generate_dlg.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/../tools/global_ids_dlg_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/../tools/preview_settings.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/artpropdlg_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/code_preference_dlg.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/dlg_gen_results.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/id_editor_dlg.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/import_winres_dlg.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/insert_widget.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menu_auibar.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menu_bartools.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menubutton.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menucheckbox.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menucombobox.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menudatactrl.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menulistbox.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menuribbontype.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menuspin.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/menustaticsizer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newdialog_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newframe_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newmdi_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newpanel_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newribbon_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/newwizard_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/optionsdlg.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/startupdlg_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/ui_images.cpp
+
+    # Base classes
+    ${CMAKE_CURRENT_LIST_DIR}/../internal/convert_img_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/../internal/msgframe_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/../tools/global_ids_dlg_base.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/artpropdlg_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/codedisplay_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/colourprop_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/dlg_gen_results.cpp
     ${CMAKE_CURRENT_LIST_DIR}/edit_html_dialog_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/editcodedialog_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/editstringdialog_base.cpp
@@ -30,32 +55,10 @@ set (wxue_generated_code
     ${CMAKE_CURRENT_LIST_DIR}/generate_xrc_dlg_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/grid_property_dlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/gridbag_item_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/id_editor_dlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/import_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/import_winres_dlg.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/insert_widget.cpp
     ${CMAKE_CURRENT_LIST_DIR}/mainframe_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menu_auibar_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menu_bartools_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menubutton_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menucheckbox_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menucombobox_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menudatactrl_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menulistbox_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menuribbontype_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menuspin_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/menustaticsizer_base.cpp
     ${CMAKE_CURRENT_LIST_DIR}/navtoolbar_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newdialog_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newframe_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newmdi_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newpanel_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newribbon_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/newwizard_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/optionsdlg.cpp
     ${CMAKE_CURRENT_LIST_DIR}/ribbonpanel_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/startupdlg_base.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/ui_images.cpp
     ${CMAKE_CURRENT_LIST_DIR}/xrccompare_base.cpp
 
 )
